Implement Halley's method for computing LN
Even if the previously implemented approximation was working, it was too slow,
because of its slow convergence. In this implementation, I use the first ~10
terms of the Taylor approximation as the starting point, and then I implement
Halley's method to achieve cubic convergence.

WARNING: this method relies on the Exp() function (which relies on the Power()
function), which at the moment is ***BROKEN*** because it only works with
integer exponents.

When the Power function is fixed to allow arbitrary exponents, then this
function should be tested.

See: https://en.wikipedia.org/wiki/Natural_logarithm#High_precision