Implement real exponent for powers
To do so, I used the identity x^y = exp(y * ln(x)). The problem was, Exp already
depended on Power (and LN depends on Exp and Power)

So I split Power in two subfunctions: one with real exponent and one with
integer exponent.

The integer exponent uses the naive implementation of iterative multiplication,
and therefore does not rely on any other function. All Taylor series that we use
in the library depend on an integer power of x for each term, so they should
work fine.

The version of Power with a real exponent uses the aforementioned identity.
Since Exp is now also based on Taylor series, it uses only integer powers.

The LN function uses either powers with integer exponent, or calls to Exp when
transitioning to Halley's method for improved precision. As a result, it should
not encounter any issue as well.