Fix Travis integration
Now that the Exp_Estimate function is no more, a bug has disappeared: rockstar-py
used to misinterpret a line in that function, and the resulting python code
was meaningless. I usually fixed it by hand, but this made CI difficult.

Since that bug is gone, I'm fixing the travis.yml file so it properly installs
the rockstar-python transpiler, and runs the series of tests

The bug was that the rockstar line:
 `While The Iterator is as little as The Number Of Terms`
was interpreted as a poetic assignment, and transpiled as:
`while the_iterator is 2623625`