Implement is_close(a,b) function
I've implemented the is_close(a,b) function as is specified in PEP 485:
https://www.python.org/dev/peps/pep-0485

At the moment there is no syntax for optional arguments, so the relative
tolerance and the absolute tolerance are hardcoded into the function code.
This will need to be changed whenever optional arguments are supported by the
language specification.